+++ src/com/intellij/idea/plugin/hybris/diagram/typeSystem/TSDiagramElementManager.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -59,14 +59,14 @@ class TSDiagramElementManager : AbstractDiagramElementManager<TSGraphNode>() {
         else -> null
     }
 
-    override fun getItemType(element: Any?) = when (element) {
-        is TSGraphFieldDeployment -> SimpleColoredText(element.value, SimpleTextAttributes.REGULAR_ATTRIBUTES)
-        is TSGraphFieldProperty -> SimpleColoredText(element.value, SimpleTextAttributes.REGULAR_ATTRIBUTES)
-        is TSGraphFieldTyped -> SimpleColoredText(element.value, SimpleTextAttributes.REGULAR_ATTRIBUTES)
-        is TSGraphFieldIndex -> SimpleColoredText(element.type, SimpleTextAttributes.REGULAR_ATTRIBUTES)
-        is TSGraphFieldAttribute -> SimpleColoredText(element.type, SimpleTextAttributes.REGULAR_ATTRIBUTES)
-        is TSGraphFieldRelationElement -> SimpleColoredText(element.type, SimpleTextAttributes.REGULAR_ATTRIBUTES)
-        is TSGraphFieldRelationEnd -> SimpleColoredText(element.type, SimpleTextAttributes.REGULAR_ATTRIBUTES)
+    override fun getItemType(nodeElement: TSGraphNode?, nodeItem: Any?, builder: DiagramBuilder?) = when (nodeItem) {
+        is TSGraphFieldDeployment -> SimpleColoredText(nodeItem.value, SimpleTextAttributes.REGULAR_ATTRIBUTES)
+        is TSGraphFieldProperty -> SimpleColoredText(nodeItem.value, SimpleTextAttributes.REGULAR_ATTRIBUTES)
+        is TSGraphFieldTyped -> SimpleColoredText(nodeItem.value, SimpleTextAttributes.REGULAR_ATTRIBUTES)
+        is TSGraphFieldIndex -> SimpleColoredText(nodeItem.type, SimpleTextAttributes.REGULAR_ATTRIBUTES)
+        is TSGraphFieldAttribute -> SimpleColoredText(nodeItem.type, SimpleTextAttributes.REGULAR_ATTRIBUTES)
+        is TSGraphFieldRelationElement -> SimpleColoredText(nodeItem.type, SimpleTextAttributes.REGULAR_ATTRIBUTES)
+        is TSGraphFieldRelationEnd -> SimpleColoredText(nodeItem.type, SimpleTextAttributes.REGULAR_ATTRIBUTES)
         else -> null
     }