+++ src/com/intellij/idea/plugin/hybris/impex/liveTemplates/ImpexContextType.kt
@@ -1,6 +1,7 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2014-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,11 +19,11 @@
 package com.intellij.idea.plugin.hybris.impex.liveTemplates
 
 import com.intellij.codeInsight.template.FileTypeBasedContextType
+import com.intellij.codeInsight.template.TemplateActionContext
 import com.intellij.idea.plugin.hybris.impex.file.ImpexFileType
 import com.intellij.idea.plugin.hybris.impex.psi.ImpexFile
-import com.intellij.psi.PsiFile
 
 class ImpexContextType : FileTypeBasedContextType("ImpEx", ImpexFileType.INSTANCE) {
 
-    override fun isInContext(file: PsiFile, offset: Int) = file is ImpexFile
+    override fun isInContext(templateActionContext: TemplateActionContext) = templateActionContext.file is ImpexFile
 }